Performance improvements:

* Update to latest version of Cranelift.
* Add support for conditional hooks with inline checks.
* Cleanup some register access APIs.

Robustness improvements:

* Better handle overflowing addresses.
* Added support for 16-bit floats.
* Handle unsigned int-to-float conversion.

General improvements and bug fixes:

* Fix issue where some zero-length strings would not be filter in comparecov instrumentation.
* Code cleanup from clippy lints.
* Update dependencies.

+++ icicle-linux/src/fs/file.rs
@@ -169,7 +169,7 @@ impl ActiveFileData {
         let new = match whence {
             types::Seek::Set => Some(offset),
             types::Seek::Cur => (self.pos as i64).checked_add(offset),
-            types::Seek::End => (self.inode.borrow().size as i64).checked_add(offset as i64),
+            types::Seek::End => (self.inode.borrow().size as i64).checked_add(offset),
             types::Seek::Data | types::Seek::Hole => return Err(errno::EINVAL),
         };
